requirements.txt
python-telegram-bot==13.0
qrcode==7.4.2
environs==9.5.0
vk-api==11.9.9
redis==4.5.5
